+++ camastersoftware.com/app/Views/super_admin/due_date/due_dates.php
@@ -304,7 +304,7 @@ if(!empty($taxCalIsAdminCookie))
                                                             $datediff = $now - $your_date;
                                                             
                                                             if($your_date>$now)
-                                                                echo abs(round($datediff / (60 * 60 * 24)));
+                                                                echo abs((float)round($datediff / (60 * 60 * 24)));
                                                             else
                                                                 echo "-";
                                                             */
